Add files via upload
This file takes an already labelled NGD_AL layer from a map in an .aprx file and converts the labels to annotations and then to points.
It creates annotations and point files by CD from a CSD_UID_L/R field.  
These points can then be made into a new map to create a VTPK of address range values.

Note that each scale needs its own layer of labels to create the annotations and points from.